add enumeration for SGFC message IDs (fixes #1)

+++ src/sgfc/frontend/SgfcDocumentReadResult.cpp
@@ -76,7 +76,7 @@ namespace LibSgfcPlusPlus
     {
       std::cout
         << "Message ID = "
-        << parseResultMessage->GetMessageID()
+        << static_cast<int>(parseResultMessage->GetMessageID())
         << ", type = "
         << (int)parseResultMessage->GetMessageType()
         << ", line = "
